+++ covalent/_dispatcher_plugins/local.py
@@ -149,14 +149,17 @@ class LocalDispatcher(BaseDispatcher):
             lattice_dispatch_id = None
             try:
                 r = requests.post(
-                    submit_dispatch_url, data=json_lattice, params={"disable_run": disable_run}
+                    submit_dispatch_url,
+                    data=json_lattice,
+                    params={"disable_run": disable_run},
+                    timeout=5,
                 )
                 r.raise_for_status()
                 lattice_dispatch_id = r.content.decode("utf-8").strip().replace('"', "")
-            except requests.exceptions.ConnectionError as e:
-                message = f"The Covalent dispatcher server is not running at {dispatcher_addr}. You must start Covalent (with `covalent start`) before dispatching your workflow."
-                app_log.error(message)
-                raise ConnectionError(message) from e
+            except requests.exceptions.ConnectionError:
+                message = f"The Covalent server cannot be reached at {dispatcher_addr}. Local servers can be started using `covalent start` in the terminal. If you are using a remote Covalent server, contact your systems administrator to report an outage."
+                print(message)
+                return
 
             if not disable_run or triggers_data is None:
                 return lattice_dispatch_id
@@ -264,8 +267,16 @@ class LocalDispatcher(BaseDispatcher):
                 dispatch_id, new_args, new_kwargs, replace_electrons, reuse_previous_results
             )
             redispatch_url = f"{dispatcher_addr}/api/redispatch"
-            r = requests.post(redispatch_url, json=body, params={"is_pending": is_pending})
-            r.raise_for_status()
+            try:
+                r = requests.post(
+                    redispatch_url, json=body, params={"is_pending": is_pending}, timeout=5
+                )
+                r.raise_for_status()
+            except requests.exceptions.ConnectionError:
+                message = f"The Covalent server cannot be reached at {dispatcher_addr}. Local servers can be started using `covalent start` in the terminal. If you are using a remote Covalent server, contact your systems administrator to report an outage."
+                print(message)
+                return
+
             return r.content.decode("utf-8").strip().replace('"', "")
 
         return func